+++ src/main/kotlin/strongdmm/event/type/service/TriggerLayersFilterService.kt
@@ -4,7 +4,7 @@ import strongdmm.event.Event
 
 abstract class TriggerLayersFilterService {
     class FilterLayersById(dmeItemIdArray: LongArray) : Event<LongArray, Unit>(dmeItemIdArray, null)
-    class ShowLayersByType(dmeItemType: String) : Event<String, Unit>(dmeItemType, null)
-    class HideLayersByType(dmeItemType: String) : Event<String, Unit>(dmeItemType, null)
+    class ShowLayersByTypeExact(dmeItemType: String) : Event<String, Unit>(dmeItemType, null)
+    class HideLayersByTypeExact(dmeItemType: String) : Event<String, Unit>(dmeItemType, null)
     class FetchFilteredLayers(callback: ((Set<String>) -> Unit)) : Event<Unit, Set<String>>(Unit, callback)
 }