+++ CHANGELOG.md
@@ -1,5 +1,6 @@
+# v1.8.0 (2020-09-13)
 Features:
- * Added an option to edit/delete instances in the "Object Panel".
+ * Added an option to globally edit/delete instances in the "Object Panel".
 
 Fixes:
  * Incorrect nudge/dir value in the tile popup after another one was opened.